test(covenantsigner): split store and server suites into thematic files
